+++ pybop/plot/problem.py
@@ -46,15 +46,15 @@ def problem(
 
     domain = problem.domain
     if problem.domain_data is None:
-        # Simulate the model for the both the initial and the given inputs
+        # Simulate the model for both the initial and the given inputs
         target = problem.target
-        problem.target = target + [domain]
-        initial_inputs = problem.simulator.parameters.to_dict("initial")
+        problem.set_target(target + [domain])
+        initial_inputs = problem.parameters.to_dict("initial")
         target_output = problem.simulate(initial_inputs)
         target_domain = target_output[domain].data
         model_output = problem.simulate(inputs)
         model_domain = model_output[domain].data
-        problem.target = target
+        problem.set_target(target)
     else:
         # Extract the time data and simulate the model for the given inputs
         target_output = Solution()
@@ -73,7 +73,7 @@ def problem(
         plot_dict = StandardPlot(
             layout_options=dict(
                 title="Scatter Plot",
-                xaxis_title="Time / s",
+                xaxis_title=StandardPlot.remove_brackets(domain),
                 yaxis_title=StandardPlot.remove_brackets(var),
             )
         )