+++ tests/unit/test_cost.py
@@ -3,10 +3,6 @@ import pybamm
 import pytest
 
 import pybop
-from pybop.costs.feature_distances import (
-    ExponentialFeatureDistance,
-    SquareRootFeatureDistance,
-)
 
 
 class TestCosts:
@@ -34,7 +30,7 @@ class TestCosts:
         return {
             "Negative electrode active material volume fraction": pybop.Parameter(
                 distribution=pybop.Gaussian(
-                    truncated_at=[0.375, 0.625], mean=0.5, sigma=0.010
+                    mean=0.5, sigma=0.010, truncated_at=[0.375, 0.625]
                 )
             )
         }
@@ -58,8 +54,14 @@ class TestCosts:
             0.2 + 0.4 * np.sqrt(domain_data[0:200]),
             0.2 + 0.4 * 20**0.5 + 3 * (2 / 3 - np.exp(-0.02 * domain_data[200:])),
         )
+        dataset = pybop.Dataset(
+            {
+                "Time [s]": domain_data,
+                "Voltage [V]": target_data,
+            }
+        )
         switchover_point = 20.0
-        return domain_data, target_data, switchover_point
+        return dataset, switchover_point
 
     def test_base(self, dataset):
         cost = pybop.ErrorMeasure(dataset)
@@ -81,14 +83,10 @@ class TestCosts:
         [
             pybop.MeanAbsoluteError,
             pybop.GaussianLogLikelihoodKnownSigma,
-            pybop.LogPosterior,
         ],
     )
     def test_fitting_costs(self, simulator, dataset, cost_class):
-        if cost_class is pybop.LogPosterior:
-            likelihood = pybop.GaussianLogLikelihoodKnownSigma(dataset, sigma=0.002)
-            cost = cost_class(likelihood)
-        elif issubclass(cost_class, pybop.LogLikelihood):
+        if issubclass(cost_class, pybop.LogLikelihood):
             cost = cost_class(dataset, sigma=0.002)
         else:
             cost = cost_class(dataset)
@@ -328,8 +326,7 @@ class TestCosts:
             {
                 "Time [s]": solution["Time [s]"].data,
                 "Current [A]": solution["Current [A]"].data,
-                "Voltage [V]": solution["Voltage [V]"].data
-                + np.random.normal(0, 0.02, len(solution["Time [s]"].data)),
+                "Voltage [V]": pybop.add_noise(solution["Voltage [V]"].data, 0.02),
             }
         )
 
@@ -405,30 +402,6 @@ class TestCosts:
                 sensitivities_2[key], sensitivities_3[key], atol=1e-5
             )
 
-        # Test LogPosterior explicitly
-        cost4 = pybop.LogPosterior(pybop.GaussianLogLikelihood(dataset))
-        weighted_cost_4 = pybop.WeightedCost(cost1, cost4, weights=[1, 1 / weight])
-        problem_4 = pybop.Problem(simulator, cost4)
-        weighted_4 = pybop.Problem(simulator, weighted_cost_4)
-        sigma = 0.01
-        assert np.isfinite(
-            cost4.parameters["Sigma for output 1"].distribution.logpdf(sigma)
-        )
-        assert np.isfinite(weighted_4([0.5, sigma]))
-        np.testing.assert_allclose(
-            weighted_4.evaluate([0.6, sigma]).values,
-            problem_1.evaluate([0.6]).values
-            - 1 / weight * problem_4.evaluate([0.6, sigma]).values,
-            atol=1e-5,
-        )
-        assert np.isfinite(weighted_4.evaluate([0.5, sigma]).values)
-        np.testing.assert_allclose(
-            weighted_4.evaluate([0.6, sigma]).values,
-            problem_1.evaluate([0.6]).values
-            - 1 / weight * problem_4.evaluate([0.6, sigma]).values,
-            atol=1e-5,
-        )
-
     def test_weighted_design_cost(self, design_simulator):
         cost_1 = pybop.DesignCost(target="Gravimetric energy density [W.h.kg-1]")
         cost_2 = pybop.DesignCost(target="Volumetric energy density [W.h.m-3]")
@@ -454,51 +427,41 @@ class TestCosts:
             pybop.WeightedCost(cost1, cost2)
 
     def test_square_root_feature_distance(self, gitt_like_dataset):
-        domain_data, target_data, switchover_point = gitt_like_dataset
-        srfd = SquareRootFeatureDistance(
-            domain_data, target_data, feature="offset", time_end=switchover_point
+        dataset, switchover_point = gitt_like_dataset
+        srfd = pybop.SquareRootFeatureDistance(
+            dataset=dataset, feature="offset", time_end=switchover_point
         )
         assert abs(srfd.data_fit - 0.2) < 1e-4
-        srfd = SquareRootFeatureDistance(
-            domain_data, target_data, feature="slope", time_end=switchover_point
+        srfd = pybop.SquareRootFeatureDistance(
+            dataset=dataset, feature="slope", time_end=switchover_point
         )
         assert abs(srfd.data_fit - 0.4) < 1e-4
-        srfd = SquareRootFeatureDistance(
-            domain_data, target_data, feature="inverse_slope", time_end=switchover_point
+        srfd = pybop.SquareRootFeatureDistance(
+            dataset=dataset, feature="inverse_slope", time_end=switchover_point
         )
         assert abs(srfd.data_fit - 1 / 0.4) < 1e-4
         with pytest.raises(ValueError):
-            srfd = SquareRootFeatureDistance(
-                domain_data, target_data, feature="non_existent"
+            srfd = pybop.SquareRootFeatureDistance(
+                dataset=dataset, feature="non_existent"
             )
 
+        assert srfd(y=np.asarray([])) == np.inf
+
     def test_exponential_feature_distance(self, gitt_like_dataset):
-        domain_data, target_data, switchover_point = gitt_like_dataset
-        efd = ExponentialFeatureDistance(
-            domain_data,
-            target_data,
-            feature="asymptote",
-            time_start=switchover_point,
+        dataset, switchover_point = gitt_like_dataset
+        efd = pybop.ExponentialFeatureDistance(
+            dataset=dataset, feature="asymptote", time_start=switchover_point
         )
         assert abs(efd.data_fit - (2.2 + 0.4 * 20**0.5)) < 1e-4
-        efd = ExponentialFeatureDistance(
-            domain_data,
-            target_data,
-            feature="magnitude",
-            time_start=switchover_point,
+        efd = pybop.ExponentialFeatureDistance(
+            dataset=dataset, feature="magnitude", time_start=switchover_point
         )
         assert abs(efd.data_fit + 2.0) < 1e-1
-        efd = ExponentialFeatureDistance(
-            domain_data,
-            target_data,
-            feature="timescale",
-            time_start=switchover_point,
+        efd = pybop.ExponentialFeatureDistance(
+            dataset=dataset, feature="timescale", time_start=switchover_point
         )
         assert abs(efd.data_fit - 1 / 0.02) < 1e-2
-        efd = ExponentialFeatureDistance(
-            domain_data,
-            target_data,
-            feature="inverse_timescale",
-            time_start=switchover_point,
+        efd = pybop.ExponentialFeatureDistance(
+            dataset=dataset, feature="inverse_timescale", time_start=switchover_point
         )
         assert abs(efd.data_fit - 0.02) < 1e-4